+++ src/Fauna/forage_distribution_algorithms.test.cpp
@@ -8,11 +8,11 @@
  * \copyright LGPL-3.0-or-later
  * \date 2019
  */
+#include "forage_distribution_algorithms.h"
 #include "catch.hpp"
 #include "dummy_herbivore.h"
 #include "dummy_hft.h"
 #include "dummy_population.h"
-#include "forage_distribution_algorithms.h"
 #include "habitat_forage.h"
 #include "hft.h"
 #include "parameters.h"