+++ src/Fauna/insfile_reader.cpp
@@ -10,6 +10,7 @@
  */
 #include "insfile_reader.h"
 #include <algorithm>
+#include <iostream>
 #include <string>
 #include "forage_types.h"
 #include "hft.h"
@@ -59,8 +60,12 @@ InsfileReader::InsfileReader(const std::string filename)
 
   {
     std::string err_msg;
-    if (!params.is_valid(err_msg))
+    const bool params_valid = params.is_valid(err_msg);
+    if (!params_valid)
       throw std::runtime_error("Parameters are not valid:\n" + err_msg);
+    else if (!err_msg.empty())
+      std::cerr << "Parameters are valid, but there were warnings:\n"
+                << err_msg;
   }
 
   if (params.herbivore_type == HerbivoreType::Cohort) {
@@ -69,9 +74,13 @@ InsfileReader::InsfileReader(const std::string filename)
       for (const auto& hft_table : *hft_table_array) {
         std::shared_ptr<const Hft> new_hft(new Hft(read_hft(hft_table)));
         std::string err_msg;
-        if (!new_hft->is_valid(params, err_msg))
+        const bool hft_valid = new_hft->is_valid(params, err_msg);
+        if (!hft_valid)
           throw std::runtime_error("HFT \"" + new_hft->name +
                                    "\" is not valid:\n" + err_msg);
+        else if (!err_msg.empty())
+          std::cerr << "HFT \"" + new_hft->name +
+                           "\" is valid, but there were warnings:\n" + err_msg;
         // Add HFT to list, but check if HFT with that name already exists.
         for (const auto& hft : hfts) {
           assert(hft.get());
@@ -166,7 +175,14 @@ void InsfileReader::check_wrong_type(
     found = "array of string";
   // Add more types here if you want to support them.
 
-  if (found != "") throw wrong_param_type(key, expected, found);
+  // Exceptions: Parameters that can be specified in different types.
+  bool exempt = false;  // Whether this parameter is an exeception.
+  if (key == "breeding_season.start") {
+    exempt = (expected == "integer" && found == "string") ||
+             (expected == "string" && found == "integer");
+  }
+
+  if (found != "" && !exempt) throw wrong_param_type(key, expected, found);
 }
 
 std::shared_ptr<cpptoml::table> InsfileReader::get_group_table(
@@ -585,10 +601,51 @@ Hft InsfileReader::read_hft(const std::shared_ptr<cpptoml::table>& table) {
     if (value) hft.breeding_season_length = *value;
   }
   {
+    const std::string key = "breeding_season.start";
     const bool mandatory = (hft.reproduction_model != ReproductionModel::None);
-    const auto value =
-        find_hft_parameter<int>(table, "breeding_season.start", mandatory);
-    if (value) hft.breeding_season_start = *value;
+    // Either an integer (Julian day) or a month name is allowed. We first
+    // check if one of them is defined with mandatory==false.
+    const auto month_name = find_hft_parameter<std::string>(table, key, false);
+    const auto julian_day = find_hft_parameter<int>(table, key, false);
+    const std::string hft_name = *table->get_as<std::string>("name");
+    if (mandatory && !month_name && !julian_day)
+      throw missing_hft_parameter(hft_name, key);
+    if (month_name && julian_day)
+      throw ambiguous_param_type(key, hft_name, *month_name,
+                                 std::to_string(*julian_day));
+    if (julian_day) hft.breeding_season_start = *julian_day;
+    // We assume 365 days in the year, for simplicity.
+    if (month_name) {
+      if (lowercase(*month_name) == "january")
+        hft.breeding_season_start = 0;
+      else if (lowercase(*month_name) == "february")
+        hft.breeding_season_start = 31;
+      else if (lowercase(*month_name) == "march")
+        hft.breeding_season_start = 59;
+      else if (lowercase(*month_name) == "april")
+        hft.breeding_season_start = 90;
+      else if (lowercase(*month_name) == "may")
+        hft.breeding_season_start = 120;
+      else if (lowercase(*month_name) == "june")
+        hft.breeding_season_start = 151;
+      else if (lowercase(*month_name) == "july")
+        hft.breeding_season_start = 181;
+      else if (lowercase(*month_name) == "august")
+        hft.breeding_season_start = 212;
+      else if (lowercase(*month_name) == "september")
+        hft.breeding_season_start = 243;
+      else if (lowercase(*month_name) == "october")
+        hft.breeding_season_start = 273;
+      else if (lowercase(*month_name) == "november")
+        hft.breeding_season_start = 304;
+      else if (lowercase(*month_name) == "december")
+        hft.breeding_season_start = 334;
+      else
+        throw invalid_option(
+            hft, key, *month_name,
+            {"January", "February", "March", "April", "May", "June", "July",
+             "August", "September", "October", "November", "December"});
+    }
   }
   {
     const bool mandatory =
@@ -835,20 +892,27 @@ void InsfileReader::read_table_output_text_tables() {
       for (const auto& s : *value)
         if (lowercase(s) == "available_forage")
           params.output_text_tables.available_forage = true;
+        else if (lowercase(s) == "body_fat")
+          params.output_text_tables.body_fat = true;
         else if (lowercase(s) == "digestibility")
           params.output_text_tables.digestibility = true;
         else if (lowercase(s) == "eaten_forage_per_ind")
           params.output_text_tables.eaten_forage_per_ind = true;
         else if (lowercase(s) == "eaten_nitrogen_per_ind")
           params.output_text_tables.eaten_nitrogen_per_ind = true;
-        else if (lowercase(s) == "mass_density_per_hft")
+        else if (lowercase(s) == "individual_density")
+          params.output_text_tables.individual_density = true;
+        else if (lowercase(s) == "mass_density")
+          params.output_text_tables.mass_density = true;
+        else if (lowercase(s) == "mass_density_per_hft")  // deprecated
           params.output_text_tables.mass_density_per_hft = true;
         // -> Add new output tables here (alphabetical order).
         else
           throw invalid_option(
               key, s,
-              {"available_forage", "digestibility", "eaten_forage_per_ind",
-               "eaten_nitrogen_per_ind", "mass_density_per_hft"});
+              {"available_forage", "body_fat", "digestibility",
+               "eaten_forage_per_ind", "eaten_nitrogen_per_ind",
+               "individual_density", "mass_density"});
     }
   }
   // Remove the table "output" in order to indicate that it’s been parsed.