+++ src/Fauna/simulation_unit.test.cpp
@@ -8,10 +8,10 @@
  * \copyright LGPL-3.0-or-later
  * \date 2019
  */
+#include "simulation_unit.h"
 #include "catch.hpp"
 #include "dummy_habitat.h"
 #include "population_list.h"
-#include "simulation_unit.h"
 using namespace Fauna;
 
 TEST_CASE("Fauna::SimulationUnit") {