+++ tests/dummy_population.test.cpp
@@ -8,8 +8,8 @@
  * \copyright LGPL-3.0-or-later
  * \date 2019
  */
-#include "catch.hpp"
 #include "dummy_population.h"
+#include "catch.hpp"
 using namespace Fauna;
 
 TEST_CASE("DummyPopulation") {